fix atari wraper for reward clip

+++ DQN/atari_wrapper.py
@@ -131,7 +131,7 @@ class ClipRewardEnv(gym.RewardWrapper):
 
     def reward(self, reward):
         """Bin reward to {+1, 0, -1} by its sign."""
-        return np.sign(reward)
+        return np.float32(np.sign(reward))
 
 
 class WarpFrame(gym.ObservationWrapper):
